fix(models): device-correct Wave 3d detection/segmentation models on Metal
All 11 Wave 3d models now run cleanly on both the CPU (Accelerate) and
Metal (MLX) compute streams. Tests parametrized with the device fixture
pass 45/45 in 61 s.

Device-propagation fixes:

* AnchorGenerator now derives device from its input feature maps and
  threads it through every internally constructed shift/anchor tensor.
* `remove_small_boxes`, `nms`, and RPN score-mask paths emit index
  tensors on the same device as their input boxes.
* `roi_align` builds its sampling grid on the input feature's device,
  and `roi_pool` returns empty crops on-device.
* `_fpn_roi_align` (Mask R-CNN) constructs all per-(b, lvl) empty
  filler tensors on-device.
* YOLO v2/v3/v4 `_decode_predictions` build col/row offset grids and
  per-anchor (w, h) maps on the device of the raw conv output.
* DETR `_build_2d_sin_pos_enc` takes a `device=` kwarg; the forward
  pass propagates `feat.device.type` to it and to the zero-initialised
  decoder target tensor. Eliminated a wasteful Python-level pos-enc
  rebuild — now expand the (S, d) encoding to (S, B, d).
* Mask2Former `_compute_mask_attn` keeps the attention-bias tensor on
  the query device.
* R-CNN / Fast R-CNN / Faster R-CNN / Mask R-CNN forward paths now
  populate every "no detections" zero tensor (logits, deltas, masks,
  proposals, boxes) on the input device.

Cross-cutting fixes uncovered while validating Metal:

* Lucid's `.clamp()` is positional-only. Replaced every `min=` /
  `max=` kwarg-only clamp in `_utils/_detection.py` with positional
  `clamp(low, high)`.
* `decode_boxes`'s `bbox_xform_clip` now uses `clamp(-1e9, clip)`.
* `Tensor.device` returns a `Device` object, not a string. Fixed
  `str(...)` mis-use by switching to `.device.type`.
* YOLOv1 loss had a single-kwarg `clamp(min=1e-6)` — now positional.

Test refactor — model objects are now built once per class/device
combination, and all assertions for a model live in a single
`test_factory_and_forward` method.  Adds `rpn_pre_nms_top_n=100` /
`num_decoder_layers=2` / `num_queries=20` overrides so Mask R-CNN,
DETR, MaskFormer, and Mask2Former finish within seconds on Metal
(NMS / Hungarian are still O(N²) Python loops with `.item()`-based
sync points, so shrinking N is essential at the unit-test scale).

+++ lucid/models/vision/efficientdet/_model.py
@@ -649,7 +649,7 @@ class EfficientDetForObjectDetection(PretrainedModel):
             cls_losses.append(_focal_loss(lg_b.reshape(-1), tgt_cls.reshape(-1)))
 
             if pos_idx:
-                pos_t = lucid.tensor(pos_idx)
+                pos_t = lucid.tensor(pos_idx).long()
                 gt_boxes_pos = lucid.tensor(
                     [
                         [float(gt_boxes[pos_gt[i], d].item()) for d in range(4)]
@@ -707,7 +707,7 @@ class EfficientDetForObjectDetection(PretrainedModel):
                 ]
                 if not mask:
                     continue
-                mask_t = lucid.tensor(mask)
+                mask_t = lucid.tensor(mask).long()
                 sc_sel = sc_c[mask_t]
                 bx_sel = bx_b[mask_t]
                 keep = batched_nms(
